// src/store/types/contactPersonTypes.ts
export interface ContactPersonFormData {
  nationality_id: string;
  country_id: string;
  id_type: string;
  title_id: string;
  telephone_code_country_id: string;
  mobile_code_country_id: string;
  first_name_arabic: string;
  last_name_arabic: string;
  role: string;
  education: string;
  date_of_birth: string;
  passport_number: string;
  passport_issue_date: string;
  passport_expiry_date: string;
  contact_person_city: string;
  address: string;
  po_box: string;
  postal_code: string;
  telephone_number: string;
  mobile_number: string;
  email: string;
  origin_contact: string;
  full_name: string;
}